full hierarchy processing
* draft API for converting the reflectance data to healpix

* include attributes from the group metadata

* allow choosing the sparse format on the gpu

* reduce the amount of rings for the kernel

* allow customizing the learning rate

* install xbatcher

* install the nvdashboard

* add pyinstrument to the profilers feature

* use the released version of rasterix

* require `healpix-geo>=0.0.10`

* import the patch-wise processing code

* allow limiting the number of patches processed

* import xbatcher

* rename

* add `device` and `format` to the parameters

* print the total number of patches

* integrate the conjugate gradient and allow choosing it

* properly integrate the conjugate gradient method

* support creating the kernel in COO format

* allow using a cartesian distance metric instead

* avoid loading data when replacing nans

* forgotten change of names

* determine neighbours twice instead of computing the internal boundary

* proper placement of replacing nans

* split the weights module into bilinear and nearest

* functions to convert lat / lon to xyz with torch

* go all-in on torch to compute the kernel

* make use of torch for the interpolation

* precompute the transpose of the weights and use CSR

* avoid a transposition error and nan values

* back to installing xdggs and healpix-geo from conda-forge

* allow passing options to the healpix grid inference

* don't require a non-existent version

* return `bool` for nearest

* require the newest `xdggs`

* support chunking the regridding matrices

* fetch the crs code from the datatree root

* function to suppress warnings

* use the source dims to extract coords

* refactor nearest neighbour interpolation to support dask

* replace the merge with partial writes and sorting after the fact

* cast the weights to the data dtype

* remove obsolete comment

* experimentation notebook for full hierarchy processing

---------

+++ legacy_converters/interpolation/matmul.py
@@ -16,7 +16,7 @@ def matmul(obj: xr.Dataset | xr.DataArray, weights: xr.DataArray, *, dims: list[
     """
 
     def _matmul_variable(var, weights):
-        sparse_result = xr.dot(weights.variable, var, dim=dims)
+        sparse_result = xr.dot(weights.variable.astype(var.dtype), var, dim=dims)
         if hasattr(sparse_result.data, "todense"):
             dense_result = sparse_result.copy(data=sparse_result.data.todense())
         else:
